Move log() to JsonApiLog class
